Added new file to appendix one
The file contains information on what might change
for the programmer when using a dependency injection container

As this is a really early version, it is not yet ready to be
reviewed. However, it will contain information on several
approaches and practices like the law of demeter, decoupling,
hollywood principle and examples how the instantiation process
changes internally and why this is a benefit.